[#234]fix: return stmt를 flow control에서 분리
