+++ app/visualize/analysis/stmt/parser/flow_control_stmt.py
@@ -4,10 +4,7 @@ from app.visualize.analysis.stmt.models.flow_control_obj import (
     BreakStmtObj,
     PassStmtObj,
     ContinueStmtObj,
-    ReturnStmtObj,
 )
-from app.visualize.analysis.stmt.parser.expr.expr_traveler import ExprTraveler
-from app.visualize.container.element_container import ElementContainer
 
 
 class BreakStmt:
@@ -26,10 +23,3 @@ class ContinueStmt:
     @staticmethod
     def parse(node: ast.Continue) -> ContinueStmtObj:
         return ContinueStmtObj(id=node.lineno)
-
-
-class ReturnStmt:
-    @staticmethod
-    def parse(node: ast.Return, elem_container: ElementContainer) -> ReturnStmtObj:
-        return_value_obj = ExprTraveler.travel(node.value, elem_container)
-        return ReturnStmtObj(id=node.lineno, value=return_value_obj.value, expr=return_value_obj.expressions)
