+++ app/visualize/analysis/stmt/parser/return_stmt.py
@@ -1,0 +1,12 @@
+import ast
+
+from app.visualize.analysis.stmt.models.flow_control_obj import ReturnStmtObj
+from app.visualize.analysis.stmt.parser.expr.expr_traveler import ExprTraveler
+from app.visualize.container.element_container import ElementContainer
+
+
+class ReturnStmt:
+    @staticmethod
+    def parse(node: ast.Return, elem_container: ElementContainer) -> ReturnStmtObj:
+        return_value_obj = ExprTraveler.travel(node.value, elem_container)
+        return ReturnStmtObj(id=node.lineno, value=return_value_obj.value, expr=return_value_obj.expressions)
